Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/media-src/media-src-blocked.sub-expected.txt

CONSOLE MESSAGE: Refused to load http://www2.localhost:8800/media/A4.ogv because it does not appear in the media-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://www2.localhost:8800/media/sound_5.oga because it does not appear in the media-src directive of the Content Security Policy.
Video element src attribute must match src list - 'none' negative test


Harness Error (TIMEOUT), message = null

PASS Disallowed async video src
PASS Disallowed async video source element
PASS Disallaowed audio src
PASS Disallowed audio source element
NOTRUN Test that securitypolicyviolation events are fired
